[MOD/#2] 폴더 구조 수정
