+++ app/src/main/java/com/sopt/now/compose/MainActivity.kt
@@ -23,6 +23,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.sopt.now.compose.feature.auth.login.LoginScreen
 import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
 
 class MainActivity : ComponentActivity() {
@@ -30,66 +31,12 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             NOWSOPTAndroidTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    LoginScreen()
-                }
+                LoginScreen()
             }
         }
     }
 }
 
-@Composable
-fun LoginScreen() {
-    var userId by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(horizontal = 40.dp)
-    ) {
-        Spacer(modifier = Modifier.padding(20.dp))
-        Text(
-            modifier = Modifier.align(Alignment.CenterHorizontally),
-            text = "Welcome To SOPT",
-            fontSize = 25.sp
-        )
-        Spacer(modifier = Modifier.weight(1f))
-        Text(text = "ID")
-        TextField(
-            value = userId,
-            onValueChange = { userId = it },
-            modifier = Modifier.fillMaxWidth(),
-            placeholder = { Text("사용자 이름 입력") }
-        )
-        Spacer(modifier = Modifier.padding(vertical = 20.dp))
-        Text("비밀번호")
-        TextField(
-            value = password,
-            onValueChange = { password = it },
-            modifier = Modifier.fillMaxWidth(),
-            placeholder = { Text("비밀번호 입력") }
-        )
-        Spacer(modifier = Modifier.weight(2f))
-        Button(
-            onClick = { /* 클릭 시 수행될 동작 */ },
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            Text("로그인 하기")
-        }
-        Spacer(modifier = Modifier.padding(vertical = 5.dp))
-        Button(
-            onClick = { /* 클릭 시 수행될 동작 */ },
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            Text("회원가입 하기")
-        }
-        Spacer(modifier = Modifier.padding(bottom = 30.dp))
-    }
-}
 
 @Preview(showBackground = true)
 @Composable
