+++ app/src/main/java/com/sopt/now/compose/feature/auth/login/LoginViewModel.kt
@@ -1,0 +1,4 @@
+package com.sopt.now.compose.feature.auth.login
+
+class LoginViewModel {
+}
