+++ app/src/main/java/com/sopt/now/compose/feature/auth/signup/SignUpScreen.kt
@@ -1,0 +1,4 @@
+package com.sopt.now.compose.feature.auth.signup
+
+class SignUpScreen {
+}
