+++ app/src/main/java/com/sopt/now/compose/feature/auth/signup/SignUpViewModel.kt
@@ -1,0 +1,4 @@
+package com.sopt.now.compose.feature.auth.signup
+
+class SignUpViewModel {
+}
