+++ app/src/main/java/com/sopt/now/compose/feature/repository/SignUpRepository.kt
@@ -1,0 +1,4 @@
+package com.sopt.now.compose.feature.repository
+
+class SignUpRepository {
+}
